python/git_testing/requirements.txt
gitpython
ipython
